requirements.txt
openpyxl
parse
parse-type
py
pylint
pytest
pytest-bdd
pytest-html
pytest-rerunfailures
pytest-sugar
pytest-xdist
python-termstyle
selenium
six
wheel
git+https://github.com/okken/pytest-expect.git@master#egg=pytest-expect
Sphinx
sphinx_rtd_theme
xlrd
request
proc
pymysql
kubernetes
rstr

